fixed hhl on fonera

+++ firewall/firewall_defines.h
@@ -34,7 +34,7 @@ struct esp_tuple
 	/* members needed for ESP protection extension */
 	uint8_t esp_prot_tfm;
 	uint32_t hash_item_length;
-	uint8_t hash_tree_depth;
+	uint32_t hash_tree_depth;
 	unsigned char *active_anchor;
 	// need for verification of anchor updates
 	unsigned char *first_active_anchor;
